requirements.txt
aiofiles==0.8.0
aioredis==1.3.1
aiosmtplib==1.1.6
aiosqlite==0.17.0
alembic==1.8.1
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
bcrypt==3.2.0
blinker==1.4
boto3==1.23.1
botocore==1.26.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.1.3
fakeredis==1.7.5
fastapi==0.75.1
fastapi-async-sqlalchemy==0.3.11
fastapi-mail==1.0.8
h11==0.12.0
hiredis==2.0.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
iniconfig==1.1.1
iso8601==1.0.2
Jinja2==3.1.1
jmespath==1.0.0
loguru==0.6.0
Mako==1.2.1
MarkupSafe==2.1.1
motor==3.0.0
packaging==21.3
passlib==1.7.4
Pillow==9.1.1
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
PyJWT==2.4.0
pymongo==4.1.1
PyMySQL==1.0.2
pyparsing==3.0.8
pypika-tortoise==0.1.4
pytest==7.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
redis==4.3.1
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
s3transfer==0.5.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.35
starlette==0.17.1
tomli==2.0.1
tortoise-orm==0.19.0
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
wrapt==1.14.1
yapf==0.32.0
